fix: address review feedback (attempt 3) — Add Tavily Search as Parallel Web-Search Option (Serper / DataForSEO)
